Modal added to confirm profile delete. Controllers updated to fix security issues by checking userId matching with request owner with targeted object'userId

+++ backend/controllers/postsCtrl.js
@@ -104,12 +104,56 @@ exports.findAll = (req, res) => {
 // ----------  DELETE  ----------  //
 exports.deletePost = (req, res, next) => {
 
-    // Soft-deletion modifying the post the ad a timestamp to deletedAt
-    Post.destroy({
-            where: {
-                id: req.params.id
+    asyncLib.waterfall([
+
+            // Checks if the request is sent from an registered user
+            function(done) {
+                User.findOne({
+                        where: { id: req.body.userId }
+                    }).then(function(userFound) {
+                        done(null, userFound);
+                    })
+                    .catch(function(err) {
+                        return res.status(500).json({ 'error': 'unable to verify user' });
+                    });
+            },
+
+            // Get the targeted post infos
+            function(userFound, done) {
+                Post.findOne({
+                        where: { id: req.params.id }
+                    })
+                    .then(function(postFound) {
+                        done(null, userFound, postFound);
+                    })
+                    .catch(function(err) {
+                        return res.status(500).json({ 'error': 'Post not found' });
+                    });
+            },
+
+            function(userFound, postFound) {
+
+                // Checks if the user is the owner of the targeted one
+                if (userFound.id == postFound.userId || userFound.isAdmin == true) { // or if he's admin
+
+                    // Soft-deletion modifying the post the ad a timestamp to deletedAt
+                    Post.destroy({
+                            where: { id: req.params.id }
+                        })
+                        .then(() => res.status(200).json({ message: 'Post supprimé !' }))
+                        .catch(error => res.status(400).json({ message: "Post introuvable", error: error }))
+
+                } else {
+                    res.status(401).json({ 'error': 'user not allowed' });
+                }
+            },
+        ],
+
+        function(userFound) {
+            if (userFound) {
+                return res.status(201).json({ 'message': 'post deleted' });
+            } else {
+                return res.status(500).json({ 'error': 'cannot delete post' });
             }
-        })
-        .then(() => res.status(200).json({ message: 'Post supprimé !' }))
-        .catch(error => res.status(400).json({ message: "Post introuvable", error: error }))
+        });
 };
